requirements.txt
requests
python-dotenv
keyring
